[paraformer] delete training relatd and simplified paraformer
