+++ wenet/paraformer/assets/global_cmvn
